updated tests, merged benchmarks into test files
this produces less visual garbage and less unnecessary files.
