+++ internal/protocol/http1/parser_test.go
@@ -1,43 +1,93 @@
 package http1
 
 import (
+	"errors"
 	"fmt"
 	"github.com/dchest/uniuri"
-	"github.com/indigo-web/indigo/internal/construct"
-	"github.com/indigo-web/indigo/internal/requestgen"
-	"github.com/indigo-web/indigo/transport/dummy"
-	"strings"
-	"testing"
-
 	"github.com/indigo-web/indigo/config"
 	"github.com/indigo-web/indigo/http"
-	"github.com/indigo-web/indigo/http/headers"
 	"github.com/indigo-web/indigo/http/method"
 	"github.com/indigo-web/indigo/http/proto"
 	"github.com/indigo-web/indigo/http/status"
+	"github.com/indigo-web/indigo/internal/construct"
+	"github.com/indigo-web/indigo/kv"
+	"github.com/indigo-web/indigo/transport/dummy"
 	"github.com/stretchr/testify/require"
+	"strings"
+	"testing"
 )
 
 func getParser(cfg *config.Config) (*parser, *http.Request) {
-	client := dummy.NewNopClient()
-	body := NewBody(client, construct.Chunked(cfg.Body), cfg.Body)
-	req := construct.Request(cfg, client, body)
-	suit := Initialize(cfg, nil, client, req, body)
+	request := construct.Request(cfg, dummy.NewNopClient())
+	keys, values, requestLine := construct.Buffers(cfg)
+	p := newParser(cfg, request, keys, values, requestLine)
 
-	return suit.parser, req
+	return p, request
+}
+
+func BenchmarkParser(b *testing.B) {
+	parser, request := getParser(config.Default())
+
+	b.Run("with 5 headers", func(b *testing.B) {
+		data := GenerateRequest(strings.Repeat("a", 500), GenerateHeaders(5))
+		b.SetBytes(int64(len(data)))
+		b.ReportAllocs()
+		b.ResetTimer()
+
+		for i := 0; i < b.N; i++ {
+			_, _, _ = parser.Parse(data)
+			request.Reset()
+		}
+	})
+
+	b.Run("with 10 headers", func(b *testing.B) {
+		data := GenerateRequest(strings.Repeat("a", 500), GenerateHeaders(10))
+		b.SetBytes(int64(len(data)))
+		b.ReportAllocs()
+		b.ResetTimer()
+
+		for i := 0; i < b.N; i++ {
+			_, _, _ = parser.Parse(data)
+			request.Reset()
+		}
+	})
+
+	b.Run("with 50 headers", func(b *testing.B) {
+		data := GenerateRequest(strings.Repeat("a", 500), GenerateHeaders(50))
+		b.SetBytes(int64(len(data)))
+		b.ReportAllocs()
+		b.ResetTimer()
+
+		for i := 0; i < b.N; i++ {
+			_, _, _ = parser.Parse(data)
+			request.Reset()
+		}
+	})
+
+	b.Run("escaped 10 headers", func(b *testing.B) {
+		data := GenerateRequest(strings.Repeat("%20", 500), GenerateHeaders(10))
+		b.SetBytes(int64(len(data)))
+		b.ReportAllocs()
+		b.ResetTimer()
+
+		for i := 0; i < b.N; i++ {
+			_, _, _ = parser.Parse(data)
+			request.Reset()
+		}
+	})
 }
 
 type wantedRequest struct {
-	Headers  headers.Headers
+	Headers  http.Headers
 	Path     string
 	Method   method.Method
-	Protocol proto.Proto
+	Protocol proto.Protocol
 }
 
 func compareRequests(t *testing.T, wanted wantedRequest, actual *http.Request) {
 	require.Equal(t, wanted.Method, actual.Method)
 	require.Equal(t, wanted.Path, actual.Path)
-	require.Equal(t, wanted.Protocol, actual.Proto)
+	require.Equal(t, wanted.Protocol, actual.Protocol)
 
 	keys := wanted.Headers.Keys()
 
@@ -59,124 +109,124 @@ func splitIntoParts(req []byte, n int) (parts [][]byte) {
 	return parts
 }
 
-func feedPartially(
-	parser *parser, rawRequest []byte, n int,
-) (state requestState, extra []byte, err error) {
-	parts := splitIntoParts(rawRequest, n)
+func feedPartially(p *parser, raw []byte, n int) (done bool, extra []byte, err error) {
+	parts := splitIntoParts(raw, n)
 
-	for _, chunk := range parts {
-		state, extra, err = parser.Parse(chunk)
+	for i, chunk := range parts {
+		done, extra, err = p.Parse(chunk)
 		if err != nil {
-			return state, extra, err
-		} else if state != Pending {
-			return state, extra, err
+			return done, extra, err
 		}
-
-		for len(extra) > 0 {
-			state, extra, err = parser.Parse(extra)
-			if state != Pending {
-				return state, extra, err
+		if done {
+			if i+1 < len(parts) {
+				return true, extra, errors.New("not all chunks were fed")
 			}
+
+			break
 		}
 	}
 
-	return state, extra, nil
+	return done, extra, err
 }
 
-func TestHttpRequestsParser_Parse_GET(t *testing.T) {
+func TestParser(t *testing.T) {
 	cfg := config.Default()
 	cfg.Headers.MaxEncodingTokens = 3
-	parser, request := getParser(cfg)
 
 	t.Run("simple GET", func(t *testing.T) {
 		raw := "GET / HTTP/1.1\r\n\r\n"
-		state, extra, err := parser.Parse([]byte(raw))
+		parser, request := getParser(cfg)
+		done, extra, err := parser.Parse([]byte(raw))
 		require.NoError(t, err)
-		require.Equal(t, HeadersCompleted, state)
+		require.True(t, done)
 		require.Empty(t, extra)
 
 		wanted := wantedRequest{
 			Method:   method.GET,
 			Path:     "/",
 			Protocol: proto.HTTP11,
-			Headers:  headers.New(),
+			Headers:  kv.New(),
 		}
 
 		compareRequests(t, wanted, request)
-		require.NoError(t, request.Reset())
+		request.Reset()
 	})
 
 	t.Run("simple GET with leading CRLF", func(t *testing.T) {
 		raw := "\r\n\r\nGET / HTTP/1.1\r\n\r\n"
-		state, extra, err := parser.Parse([]byte(raw))
+		parser, request := getParser(cfg)
+		done, extra, err := parser.Parse([]byte(raw))
 		require.Error(t, err, status.ErrBadRequest.Error())
-		require.Equal(t, Error, state)
+		require.True(t, done)
 		require.Empty(t, extra)
-
-		require.NoError(t, request.Reset())
+		request.Reset()
 	})
 
 	t.Run("normal GET", func(t *testing.T) {
 		raw := "GET / HTTP/1.1\r\nHello: World!\r\nEaster: Egg\r\n\r\n"
-		state, extra, err := parser.Parse([]byte(raw))
+		parser, request := getParser(cfg)
+		done, extra, err := parser.Parse([]byte(raw))
 		require.NoError(t, err)
-		require.Equal(t, HeadersCompleted, state)
+		require.True(t, done)
 		require.Empty(t, extra)
 
 		wanted := wantedRequest{
 			Method:   method.GET,
 			Path:     "/",
 			Protocol: proto.HTTP11,
-			Headers: headers.NewFromMap(map[string][]string{
+			Headers: kv.NewFromMap(map[string][]string{
 				"hello": {"World!"},
 			}),
 		}
 
 		compareRequests(t, wanted, request)
-		require.NoError(t, request.Reset())
+		request.Reset()
 	})
 
 	t.Run("multiple header values", func(t *testing.T) {
 		raw := "GET / HTTP/1.1\r\nAccept: one,two\r\nAccept: three\r\n\r\n"
-		state, extra, err := parser.Parse([]byte(raw))
+		parser, request := getParser(cfg)
+		done, extra, err := parser.Parse([]byte(raw))
 		require.NoError(t, err)
-		require.Equal(t, HeadersCompleted, state)
+		require.True(t, done)
 		require.Empty(t, extra)
 
 		wanted := wantedRequest{
 			Method:   method.GET,
 			Path:     "/",
 			Protocol: proto.HTTP11,
-			Headers: headers.NewFromMap(map[string][]string{
+			Headers: kv.NewFromMap(map[string][]string{
 				"accept": {"one,two", "three"},
 			}),
 		}
 
 		compareRequests(t, wanted, request)
-		require.NoError(t, request.Reset())
+		request.Reset()
 	})
 
 	t.Run("only lf", func(t *testing.T) {
 		raw := "GET / HTTP/1.1\nHello: World!\n\n"
-		state, extra, err := parser.Parse([]byte(raw))
+		parser, request := getParser(cfg)
+		done, extra, err := parser.Parse([]byte(raw))
 		require.NoError(t, err)
-		require.Equal(t, HeadersCompleted, state)
+		require.True(t, done)
 		require.Empty(t, extra)
 
 		wanted := wantedRequest{
 			Method:   method.GET,
 			Path:     "/",
 			Protocol: proto.HTTP11,
-			Headers: headers.NewFromMap(map[string][]string{
+			Headers: kv.NewFromMap(map[string][]string{
 				"hello": {"World!"},
 			}),
 		}
 
 		compareRequests(t, wanted, request)
-		require.NoError(t, request.Reset())
+		request.Reset()
 	})
 
 	t.Run("escaping", func(t *testing.T) {
+		parser, request := getParser(cfg)
 		tcs := []struct {
 			Raw      []byte
 			WantPath string
@@ -186,254 +236,242 @@ func TestHttpRequestsParser_Parse_GET(t *testing.T) {
 				"/hello, world",
 			},
 			{
-				requestgen.Generate(strings.Repeat("%20", 500), requestgen.Headers(10)),
+				GenerateRequest(strings.Repeat("%20", 500), GenerateHeaders(10)),
 				"/" + strings.Repeat(" ", 500),
 			},
 		}
 
 		for i, tc := range tcs {
-			state, extra, err := parser.Parse(tc.Raw)
+			done, extra, err := parser.Parse(tc.Raw)
 			require.NoError(t, err, i)
-			require.Equal(t, HeadersCompleted, state, i)
+			require.True(t, done, i)
 			require.Empty(t, extra, i)
 
 			wanted := wantedRequest{
 				Method:   method.GET,
 				Path:     tc.WantPath,
 				Protocol: proto.HTTP11,
-				Headers:  headers.New(),
+				Headers:  kv.New(),
 			}
 
 			compareRequests(t, wanted, request)
-			require.NoError(t, request.Reset())
+			request.Reset()
 		}
 
 	})
 
 	t.Run("fuzz GET", func(t *testing.T) {
 		raw := "GET / HTTP/1.1\r\nHello: World!\r\nEaster: Egg\r\n\r\n"
+		parser, request := getParser(cfg)
 
 		for i := 1; i < len(raw); i++ {
-			state, extra, err := feedPartially(parser, []byte(raw), i)
+			done, extra, err := feedPartially(parser, []byte(raw), i)
 			require.NoError(t, err, i)
 			require.Empty(t, extra)
-			require.Equal(t, HeadersCompleted, state)
+			require.True(t, done)
 
 			wanted := wantedRequest{
 				Method:   method.GET,
 				Path:     "/",
 				Protocol: proto.HTTP11,
-				Headers: headers.NewFromMap(map[string][]string{
+				Headers: kv.NewFromMap(map[string][]string{
 					"hello": {"World!"},
 				}),
 			}
 
 			compareRequests(t, wanted, request)
-			require.NoError(t, request.Reset())
+			request.Reset()
 		}
 	})
 
 	t.Run("absolute path", func(t *testing.T) {
 		raw := "GET http://www.w3.org/pub/WWW/TheProject.html HTTP/1.1\r\n\r\n"
-		state, extra, err := parser.Parse([]byte(raw))
+		parser, request := getParser(cfg)
+		done, extra, err := parser.Parse([]byte(raw))
 		require.NoError(t, err)
-		require.Equal(t, HeadersCompleted, state)
+		require.True(t, done)
 		require.Empty(t, extra)
 
 		wanted := wantedRequest{
 			Method:   method.GET,
 			Path:     "http://www.w3.org/pub/WWW/TheProject.html",
 			Protocol: proto.HTTP11,
-			Headers:  headers.New(),
+			Headers:  kv.New(),
 		}
 
 		compareRequests(t, wanted, request)
-		require.NoError(t, request.Reset())
+		request.Reset()
 	})
 
 	t.Run("content-length", func(t *testing.T) {
 		raw := "GET / HTTP/1.1\r\nContent-Length: 13\n\r\nHello, world!"
-		state, extra, err := parser.Parse([]byte(raw))
+		parser, request := getParser(cfg)
+		done, extra, err := parser.Parse([]byte(raw))
 		require.NoError(t, err)
-		require.Equal(t, HeadersCompleted, state)
+		require.True(t, done)
 		require.Equal(t, "Hello, world!", string(extra))
 		require.Equal(t, 13, request.ContentLength)
-		require.NoError(t, request.Reset())
+		request.Reset()
 
 		raw = "GET / HTTP/1.1\r\nContent-Length: 13\r\nHi-Hi: ha-ha\r\n\r\nHello, world!"
-		state, extra, err = parser.Parse([]byte(raw))
+		done, extra, err = parser.Parse([]byte(raw))
 		require.NoError(t, err)
-		require.Equal(t, HeadersCompleted, state)
+		require.True(t, done)
 		require.Equal(t, "Hello, world!", string(extra))
 		require.Equal(t, 13, request.ContentLength)
 		require.True(t, request.Headers.Has("hi-hi"))
 		require.Equal(t, "ha-ha", request.Headers.Value("hi-hi"))
-		require.NoError(t, request.Reset())
+		request.Reset()
 	})
 
 	t.Run("connection", func(t *testing.T) {
 		raw := "GET / HTTP/1.1\r\nConnection: Keep-Alive\r\n\r\n"
-		state, extra, err := parser.Parse([]byte(raw))
+		parser, request := getParser(cfg)
+		done, extra, err := parser.Parse([]byte(raw))
 		require.NoError(t, err)
-		require.Equal(t, HeadersCompleted, state)
+		require.True(t, done)
 		require.Empty(t, string(extra))
 		require.Equal(t, "Keep-Alive", request.Connection)
-		require.NoError(t, request.Reset())
+		request.Reset()
 	})
 
 	t.Run("Transfer-Encoding and Content-Encoding", func(t *testing.T) {
-		raw := "GET / HTTP/1.1\r\nTransfer-Encoding: gzip, deflate, chunked\r\nContent-Encoding: png, jpeg, gz\r\n\r\n"
-		state, extra, err := parser.Parse([]byte(raw))
+		raw := "GET / HTTP/1.1\r\nTransfer-Encoding: gzip, deflate, chunked\r\nContent-Encoding: png, chunked, gz\r\n\r\n"
+		parser, request := getParser(cfg)
+		done, extra, err := parser.Parse([]byte(raw))
 		require.NoError(t, err)
-		require.Equal(t, HeadersCompleted, state)
+		require.True(t, done)
 		require.Empty(t, string(extra))
-		require.Equal(t, []string{"gzip", "deflate", "chunked"}, request.Encoding.Transfer)
+		require.Equal(t, []string{"gzip", "deflate"}, request.Encoding.Transfer)
 		require.True(t, request.Encoding.Chunked)
-		require.Equal(t, []string{"png", "jpeg", "gz"}, request.Encoding.Content)
-		require.NoError(t, request.Reset())
-	})
-}
-
-func TestHttpRequestsParser_POST(t *testing.T) {
-	parser, request := getParser(config.Default())
-
-	t.Run("fuzz POST by different chunk sizes", func(t *testing.T) {
-		raw := "POST / HTTP/1.1\r\nHello: World!\r\nContent-Length: 13\r\n\r\nHello, World!"
-
-		for i := 1; i < len(raw); i++ {
-			state, _, err := feedPartially(parser, []byte(raw), i)
-			require.NoError(t, err)
-			require.Equal(t, HeadersCompleted, state)
-
-			wanted := wantedRequest{
-				Method:   method.POST,
-				Path:     "/",
-				Protocol: proto.HTTP11,
-				Headers: headers.NewFromMap(map[string][]string{
-					"hello": {"World!"},
-				}),
-			}
-
-			compareRequests(t, wanted, request)
-			require.NoError(t, request.Reset())
-		}
+		require.Equal(t, []string{"png", "chunked", "gz"}, request.Encoding.Content)
+		request.Reset()
 	})
 
-	t.Run("query in a path", func(t *testing.T) {
+	t.Run("path params", func(t *testing.T) {
 		raw := "GET /path?hello=world HTTP/1.1\r\n\r\n"
-		state, extra, err := parser.Parse([]byte(raw))
+		parser, request := getParser(cfg)
+		done, extra, err := parser.Parse([]byte(raw))
 		require.NoError(t, err)
-		require.Equal(t, HeadersCompleted, state)
+		require.True(t, done)
 		require.Empty(t, extra)
 
 		wanted := wantedRequest{
 			Method:   method.GET,
 			Path:     "/path",
 			Protocol: proto.HTTP11,
-			Headers:  headers.New(),
+			Headers:  kv.New(),
 		}
 
 		compareRequests(t, wanted, request)
-		require.Equal(t, "hello=world", request.Query.String())
-		require.NoError(t, request.Reset())
+		require.Equal(t, "world", request.Params.Value("hello"))
+		request.Reset()
 	})
 }
 
-func TestHttpRequestsParser_Parse_Negative(t *testing.T) {
+func TestParser_Edgecases(t *testing.T) {
 	t.Run("no method", func(t *testing.T) {
 		parser, _ := getParser(config.Default())
 		raw := []byte(" / HTTP/1.1\r\n\r\n")
-		state, _, err := parser.Parse(raw)
+		done, _, err := parser.Parse(raw)
 		require.EqualError(t, err, status.ErrBadRequest.Error())
-		require.Equal(t, Error, state)
+		require.True(t, done)
 	})
 
 	t.Run("no path", func(t *testing.T) {
 		parser, _ := getParser(config.Default())
 		raw := []byte("GET HTTP/1.1\r\n\r\n")
-		state, _, err := parser.Parse(raw)
+		done, _, err := parser.Parse(raw)
 		require.EqualError(t, err, status.ErrBadRequest.Error())
-		require.Equal(t, Error, state)
+		require.True(t, done)
 	})
 
 	t.Run("whitespace as a path", func(t *testing.T) {
 		parser, _ := getParser(config.Default())
 		raw := []byte("GET  HTTP/1.1\r\n\r\n")
-		state, _, err := parser.Parse(raw)
+		done, _, err := parser.Parse(raw)
 		require.EqualError(t, err, status.ErrBadRequest.Error())
-		require.Equal(t, Error, state)
+		require.True(t, done)
 	})
 
 	t.Run("short invalid method", func(t *testing.T) {
 		parser, _ := getParser(config.Default())
 		raw := []byte("GE / HTTP/1.1\r\n\r\n")
-		state, _, err := parser.Parse(raw)
+		done, _, err := parser.Parse(raw)
 		require.EqualError(t, err, status.ErrMethodNotImplemented.Error())
-		require.Equal(t, Error, state)
+		require.True(t, done)
+	})
+
+	t.Run("normal invalid method", func(t *testing.T) {
+		parser, _ := getParser(config.Default())
+		raw := []byte("GOT / HTTP/1.1\r\n\r\n")
+		done, _, err := parser.Parse(raw)
+		require.EqualError(t, err, status.ErrMethodNotImplemented.Error())
+		require.True(t, done)
 	})
 
 	t.Run("long invalid method", func(t *testing.T) {
 		parser, _ := getParser(config.Default())
 		raw := []byte("PATCHPOSTPUT / HTTP/1.1\r\n\r\n")
-		state, _, err := parser.Parse(raw)
+		done, _, err := parser.Parse(raw)
 		require.EqualError(t, err, status.ErrMethodNotImplemented.Error())
-		require.Equal(t, Error, state)
+		require.True(t, done)
 	})
 
 	t.Run("short invalid protocol", func(t *testing.T) {
 		parser, _ := getParser(config.Default())
 		raw := []byte("GET / HTT\r\n\r\n")
-		state, _, err := parser.Parse(raw)
-		require.EqualError(t, err, status.ErrUnsupportedProtocol.Error())
-		require.Equal(t, Error, state)
+		done, _, err := parser.Parse(raw)
+		require.EqualError(t, err, status.ErrHTTPVersionNotSupported.Error())
+		require.True(t, done)
 	})
 
 	t.Run("long invalid protocol", func(t *testing.T) {
 		parser, _ := getParser(config.Default())
 		raw := []byte("GET / HTTPS/1.1\r\n\r\n")
-		state, _, err := parser.Parse(raw)
-		require.EqualError(t, err, status.ErrUnsupportedProtocol.Error())
-		require.Equal(t, Error, state)
+		done, _, err := parser.Parse(raw)
+		require.EqualError(t, err, status.ErrHTTPVersionNotSupported.Error())
+		require.True(t, done)
 	})
 
 	t.Run("unsupported minor version", func(t *testing.T) {
 		parser, _ := getParser(config.Default())
 		raw := []byte("GET / HTTP/1.2\r\n\r\n")
-		state, _, err := parser.Parse(raw)
-		require.EqualError(t, err, status.ErrUnsupportedProtocol.Error())
-		require.Equal(t, Error, state)
+		done, _, err := parser.Parse(raw)
+		require.EqualError(t, err, status.ErrHTTPVersionNotSupported.Error())
+		require.True(t, done)
 	})
 
 	t.Run("unsupported major version", func(t *testing.T) {
 		parser, _ := getParser(config.Default())
 		raw := []byte("GET / HTTP/42.1\r\n\r\n")
-		state, _, err := parser.Parse(raw)
-		require.EqualError(t, err, status.ErrUnsupportedProtocol.Error())
-		require.Equal(t, Error, state)
+		done, _, err := parser.Parse(raw)
+		require.EqualError(t, err, status.ErrHTTPVersionNotSupported.Error())
+		require.True(t, done)
 	})
 
 	t.Run("invalid minor version", func(t *testing.T) {
 		parser, _ := getParser(config.Default())
 		raw := []byte("GET / HTTP/1.x\r\n\r\n")
-		state, _, err := parser.Parse(raw)
-		require.EqualError(t, err, status.ErrUnsupportedProtocol.Error())
-		require.Equal(t, Error, state)
+		done, _, err := parser.Parse(raw)
+		require.EqualError(t, err, status.ErrHTTPVersionNotSupported.Error())
+		require.True(t, done)
 	})
 
 	t.Run("invalid major version", func(t *testing.T) {
 		parser, _ := getParser(config.Default())
 		raw := []byte("GET / HTTP/x.1\r\n\r\n")
-		state, _, err := parser.Parse(raw)
-		require.EqualError(t, err, status.ErrUnsupportedProtocol.Error())
-		require.Equal(t, Error, state)
+		done, _, err := parser.Parse(raw)
+		require.EqualError(t, err, status.ErrHTTPVersionNotSupported.Error())
+		require.True(t, done)
 	})
 
 	t.Run("lfcr crlf break sequence", func(t *testing.T) {
 		parser, _ := getParser(config.Default())
 		raw := []byte("GET / HTTP/1.1\n\r\r\n")
-		state, _, err := parser.Parse(raw)
+		done, _, err := parser.Parse(raw)
 		require.EqualError(t, err, status.ErrBadRequest.Error())
-		require.Equal(t, Error, state)
+		require.True(t, done)
 	})
 
 	t.Run("lfcr lfcr break sequence", func(t *testing.T) {
@@ -442,10 +480,10 @@ func TestHttpRequestsParser_Parse_Negative(t *testing.T) {
 		// the last one CR will be returned as extra-bytes
 		parser, _ := getParser(config.Default())
 		raw := []byte("GET / HTTP/1.1\n\r\n\r")
-		state, extra, err := parser.Parse(raw)
+		done, extra, err := parser.Parse(raw)
 		require.Equal(t, []byte("\r"), extra)
 		require.NoError(t, err)
-		require.Equal(t, HeadersCompleted, state)
+		require.True(t, done)
 	})
 
 	t.Run("invalid content length", func(t *testing.T) {
@@ -461,9 +499,9 @@ func TestHttpRequestsParser_Parse_Negative(t *testing.T) {
 		// something like a deprecated mechanism
 		parser, _ := getParser(config.Default())
 		raw := []byte("GET / \r\n")
-		state, _, err := parser.Parse(raw)
-		require.EqualError(t, err, status.ErrUnsupportedProtocol.Error())
-		require.Equal(t, Error, state)
+		done, _, err := parser.Parse(raw)
+		require.EqualError(t, err, status.ErrHTTPVersionNotSupported.Error())
+		require.True(t, done)
 	})
 
 	t.Run("too long header key", func(t *testing.T) {
@@ -501,8 +539,8 @@ func TestHttpRequestsParser_Parse_Negative(t *testing.T) {
 	t.Run("duplicate chunked TE token", func(t *testing.T) {
 		parser, _ := getParser(config.Default())
 		raw := "GET / HTTP/1.1\r\nTransfer-Encoding: gzip, deflate, chunked, chunked\r\n\r\n"
-		state, extra, err := parser.Parse([]byte(raw))
-		require.Equal(t, Error, state)
+		done, extra, err := parser.Parse([]byte(raw))
+		require.True(t, done)
 		require.Empty(t, string(extra))
 		require.EqualError(t, err, status.ErrBadEncoding.Error())
 	})
@@ -512,8 +550,8 @@ func TestHttpRequestsParser_Parse_Negative(t *testing.T) {
 		cfg.Headers.MaxEncodingTokens = 3
 		parser, _ := getParser(cfg)
 		raw := "GET / HTTP/1.1\r\nTransfer-Encoding: gzip, deflate, br, chunked\r\n\r\n"
-		state, extra, err := parser.Parse([]byte(raw))
-		require.Equal(t, Error, state)
+		done, extra, err := parser.Parse([]byte(raw))
+		require.True(t, done)
 		require.Empty(t, string(extra))
 		require.EqualError(t, err, status.ErrTooManyEncodingTokens.Error())
 	})
